requirements.txt
flask==2.0.2
Werkzeug==2.0.3
jinja2==3.0.2
flask_login==0.5.0
flask_migrate==3.1.0
WTForms==3.0.0
flask_wtf==1.0.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.29
email_validator==1.1.3
gunicorn==20.1.0
flask-restx==0.5.1
python-dotenv==0.19.2
Flask-Minify==0.37
azure-common==1.1.28
azure-core==1.25.1
azure-identity==1.11.0
azure-keyvault-secrets==4.6.0
azure-mgmt-core==1.3.2
azure-mgmt-rdbms==10.1.0
bcrypt==4.0.1
cachelib==0.9.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
Flask-Bcrypt==1.0.1
Flask-OAuthlib==0.9.6
greenlet==1.1.3
idna==3.4
importlib-metadata==5.1.0
isodate==0.6.1
itsdangerous==2.1.2
MarkupSafe==2.1.1
msal==1.19.0
msal-extensions==1.0.0
msrest==0.7.1
nose==1.3.7
oauthlib==2.1.0
portalocker==2.5.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.5.0
requests==2.28.1
requests-oauthlib==1.1.0
six==1.16.0
SQLAlchemy-Utils==0.38.3
tqdm==4.64.1
types-cryptography==3.3.23
typing-extensions==4.3.0
urllib3==1.26.12
zipp==3.11.0

# flask_mysqldb
# psycopg2-binary
